cluster: fix compiling errors caused by git rebase
