+++ server/handler.go
@@ -817,11 +817,11 @@ func (h *Handler) ResetTS(ts uint64) error {
 // SetStoreLimitScene sets the limit values for differents scenes
 func (h *Handler) SetStoreLimitScene(scene *schedule.StoreLimitScene) {
 	cluster := h.s.GetRaftCluster()
-	cluster.limiter.ReplaceStoreLimitScene(scene)
+	cluster.GetStoreLimiter().ReplaceStoreLimitScene(scene)
 }
 
 // GetStoreLimitScene returns the limit valus for different scenes
 func (h *Handler) GetStoreLimitScene() *schedule.StoreLimitScene {
 	cluster := h.s.GetRaftCluster()
-	return cluster.limiter.StoreLimitScene()
+	return cluster.GetStoreLimiter().StoreLimitScene()
 }
